Register fiber monitor MBean outside of constructor
